Create moves.cpp
Hello

This is the code to equalise all the elements in an array. for eg user has entered an array but it contains different elements or numbers. So by this code we can make all elements of the array equal. It will print the minimum number of moves required to make all elements of the array equal. 

Thank you.